Redo lexing to fix some bugs
Before this, the input "f(x" produced `[Label("f"), Label("(x")]` rather than
`[Label("f"), OpenParen, Label("x")]`.  This redoes the lexing system and has to
redo the `TaggedIter` class minorly by converting `next()` to `advance()`.  This
simplifies this class as it separates the concerns of if there is a next
character from recognizing it.

Note that the actual representation isn't exactly like this, but for the
purposes of readability this notation is useful.